factor out the new unix stream server to a new module
