+++ c8ylp/tcp_socket/tcp_server.py
@@ -17,7 +17,6 @@
 #
 """TCP server"""
 
-import os
 import errno
 import logging
 import socket
@@ -120,29 +119,6 @@ class CustomTCPServer(socketserver.TCPServer):
         )
 
 
-class CustomUnixStreamServer(socketserver.UnixStreamServer):
-    """Custom TCP Server used to listen for local connections and proxy them to
-    a websocket
-    """
-
-    def __init__(
-        self,
-        web_socket_client,
-        server_address,
-        RequestHandlerClass,
-        bind_and_activate=True,
-        buffer_size: int = 1024,
-        tcp_timeout: float = 30,
-    ) -> None:
-        self.allow_reuse_address = True
-        self.timeout = tcp_timeout
-        self.buffer_size = buffer_size
-        self.web_socket_client = web_socket_client
-        super().__init__(
-            server_address, RequestHandlerClass, bind_and_activate=bind_and_activate
-        )
-
-
 class TCPProxyServer:
     """TCP Server"""
 
@@ -198,38 +174,3 @@ class TCPProxyServer:
             bool: True if the server is running
         """
         return self._running.wait(timeout)
-
-
-class UnixStreamProxyServer(TCPProxyServer):
-    """
-    Unix domain socket server
-    """
-
-    # pylint: disable=super-init-not-called
-    def __init__(
-        self,
-        path,
-        web_socket_client,
-        tcp_buffer_size,
-        tcp_timeout,
-    ):
-        self.web_socket_client = web_socket_client
-        self._running = threading.Event()
-        self.logger = logging.getLogger(__name__)
-        self.path = path
-
-        # Expose func to web socket client
-        self.web_socket_client.proxy = self
-
-        self.server = CustomUnixStreamServer(
-            self.web_socket_client,
-            path,
-            TCPHandler,
-            buffer_size=tcp_buffer_size,
-            tcp_timeout=tcp_timeout,
-        )
-
-    def shutdown(self):
-        """Shutdown tcp server"""
-        super().shutdown()
-        os.remove(self.path)
